Move the checkout phase from the installation to the plan screen to allow checkout tracking
