+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/login/storecreation/plans/PlansFragment.kt
@@ -10,15 +10,21 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.woocommerce.android.extensions.navigateSafely
 import com.woocommerce.android.ui.base.BaseFragment
+import com.woocommerce.android.ui.common.wpcomwebview.WPComWebViewAuthenticator
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
 import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.NavigateToNextStep
 import com.woocommerce.android.ui.main.AppBarStatus
 import com.woocommerce.android.viewmodel.MultiLiveEvent
 import dagger.hilt.android.AndroidEntryPoint
+import org.wordpress.android.fluxc.network.UserAgent
+import javax.inject.Inject
 
 @AndroidEntryPoint
 class PlansFragment : BaseFragment() {
     private val viewModel: PlansViewModel by viewModels()
+
+    @Inject internal lateinit var authenticator: WPComWebViewAuthenticator
+    @Inject internal lateinit var userAgent: UserAgent
 
     override val activityAppBarStatus: AppBarStatus
         get() = AppBarStatus.Hidden
@@ -28,7 +34,7 @@ class PlansFragment : BaseFragment() {
             setViewCompositionStrategy(ViewCompositionStrategy.DisposeOnViewTreeLifecycleDestroyed)
             setContent {
                 WooThemeWithBackground {
-                    PlanScreen(viewModel = viewModel)
+                    PlanScreen(viewModel = viewModel, authenticator, userAgent)
                 }
             }
         }
