+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/login/storecreation/plans/PlansViewModel.kt
@@ -7,11 +7,22 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.asLiveData
 import com.woocommerce.android.R.drawable
 import com.woocommerce.android.R.string
-import com.woocommerce.android.analytics.AnalyticsEvent
+import com.woocommerce.android.analytics.AnalyticsEvent.SITE_CREATION_STEP
 import com.woocommerce.android.analytics.AnalyticsTracker
+import com.woocommerce.android.analytics.AnalyticsTracker.Companion.VALUE_STEP_PLAN_PURCHASE
+import com.woocommerce.android.analytics.AnalyticsTracker.Companion.VALUE_STEP_WEB_CHECKOUT
 import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
-import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.Plan.BillingPeriod.MONTHLY
-import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.Plan.Feature
+import com.woocommerce.android.ui.login.storecreation.NewStore
+import com.woocommerce.android.ui.login.storecreation.StoreCreationRepository
+import com.woocommerce.android.ui.login.storecreation.StoreCreationRepository.SiteCreationData
+import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.PlanInfo.BillingPeriod.MONTHLY
+import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.PlanInfo.BillingPeriod.YEARLY
+import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.PlanInfo.Feature
+import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.ViewState.CheckoutState
+import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.ViewState.ErrorState
+import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.ViewState.ErrorState.PlanPurchaseError
+import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.ViewState.ErrorState.PlanPurchaseError.PLAN_PURCHASE_FAILED
+import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.ViewState.ErrorState.PlanPurchaseError.SITE_CREATION_FAILED
 import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.ViewState.LoadingState
 import com.woocommerce.android.ui.login.storecreation.plans.PlansViewModel.ViewState.PlanState
 import com.woocommerce.android.viewmodel.MultiLiveEvent
@@ -22,14 +33,23 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.parcelize.Parcelize
+import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooResult
+import java.util.TimeZone
 import javax.inject.Inject
 
 @HiltViewModel
 class PlansViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
-    val analyticsTrackerWrapper: AnalyticsTrackerWrapper
+    private val analyticsTrackerWrapper: AnalyticsTrackerWrapper,
+    private val newStore: NewStore,
+    private val repository: StoreCreationRepository
 ) : ScopedViewModel(savedStateHandle) {
     companion object {
+        private const val NEW_SITE_LANGUAGE_ID = "en"
+        private const val NEW_SITE_THEME = "pub/zoologist"
+        private const val CART_URL = "https://wordpress.com/checkout"
+        private const val WEBVIEW_SUCCESS_TRIGGER_KEYWORD = "https://wordpress.com/checkout/thank-you/"
+        private const val WEBVIEW_EXIT_TRIGGER_KEYWORD = "https://woocommerce.com/"
         private const val ECOMMERCE_PLAN_NAME = "eCommerce"
         private const val ECOMMERCE_PLAN_PRICE_MONTHLY = "$70"
     }
@@ -39,10 +59,14 @@ class PlansViewModel @Inject constructor(
 
     init {
         loadPlan()
+        trackStep(VALUE_STEP_PLAN_PURCHASE)
+    }
+
+    private fun trackStep(step: String) {
         analyticsTrackerWrapper.track(
-            AnalyticsEvent.SITE_CREATION_STEP,
+            SITE_CREATION_STEP,
             mapOf(
-                AnalyticsTracker.KEY_STEP to AnalyticsTracker.VALUE_STEP_PLAN_PURCHASE
+                AnalyticsTracker.KEY_STEP to step
             )
         )
     }
@@ -91,18 +115,74 @@ class PlansViewModel @Inject constructor(
         }
     }
 
-    fun onCloseClicked() {
+    private fun showCheckoutWebsite() {
+        _viewState.update {
+            CheckoutState(
+                startUrl = "${CART_URL}/${newStore.data.domain}",
+                successTriggerKeyword = WEBVIEW_SUCCESS_TRIGGER_KEYWORD,
+                exitTriggerKeyword = WEBVIEW_EXIT_TRIGGER_KEYWORD
+            )
+        }
+        trackStep(VALUE_STEP_WEB_CHECKOUT)
+    }
+
+    private fun launchInstallation() {
+        suspend fun <T : Any?> handleErrorOrProceed(
+            result: WooResult<T>,
+            error: PlanPurchaseError,
+            successAction: suspend () -> Unit
+        ) {
+            when {
+                result.isError -> {
+                    _viewState.update { ErrorState(error, result.error.message) }
+                }
+                result.model == null -> _viewState.update { ErrorState(error) }
+                else -> successAction()
+            }
+        }
+
+        _viewState.update { LoadingState }
+
+        launch {
+            val siteCreationResult = createSite()
+            handleErrorOrProceed(siteCreationResult, SITE_CREATION_FAILED) {
+                val cartCreationResult = repository.addPlanToCart(newStore.data.siteId!!)
+                handleErrorOrProceed(cartCreationResult, PLAN_PURCHASE_FAILED) {
+                    showCheckoutWebsite()
+                }
+            }
+        }
+    }
+
+    private suspend fun createSite(): WooResult<Long> {
+        val result = repository.createNewSite(
+            SiteCreationData(
+                siteDesign = NEW_SITE_THEME,
+                domain = newStore.data.domain,
+                title = newStore.data.name,
+                segmentId = null
+            ),
+            NEW_SITE_LANGUAGE_ID,
+            TimeZone.getDefault().id
+        )
+
+        if (!result.isError) {
+            newStore.update(siteId = result.model)
+        }
+
+        return result
+    }
+
+    fun onExitTriggered() {
         triggerEvent(Exit)
     }
 
     fun onConfirmClicked() {
-        triggerEvent(NavigateToNextStep)
-        // TODO add tracking for login_woocommerce_site_created and site_creation_failed once we have the result
-        // TODO wipe out data from "NewStore" singlenton once store is created successfully
+        launchInstallation()
     }
 
-    fun onRetryClicked() {
-        // TODO
+    fun onStoreCreated() {
+        triggerEvent(NavigateToNextStep)
     }
 
     sealed interface ViewState : Parcelable {
@@ -110,16 +190,28 @@ class PlansViewModel @Inject constructor(
         object LoadingState : ViewState
 
         @Parcelize
-        object ErrorState : ViewState
+        data class ErrorState(val error: PlanPurchaseError, val message: String? = null) : ViewState {
+            enum class PlanPurchaseError {
+                SITE_CREATION_FAILED,
+                PLAN_PURCHASE_FAILED
+            }
+        }
+
+        @Parcelize
+        data class CheckoutState(
+            val startUrl: String,
+            val successTriggerKeyword: String,
+            val exitTriggerKeyword: String
+        ) : ViewState
 
         @Parcelize
         data class PlanState(
-            val plan: Plan
+            val plan: PlanInfo
         ) : ViewState
     }
 
     @Parcelize
-    data class Plan(
+    data class PlanInfo(
         val name: String,
         val billingPeriod: BillingPeriod,
         val formattedPrice: String,
